CMS-98 Accounts / Memberships graph: nodes has meaningless labels
Seems like something has gone wrong with the memberships graph now.
When checking the memberships for a user I see a lot lot of nodes with meaningless labels.

Not sure if these nodes are memeberships or something is wrong with the data.

See the attached screen shot for better explanation.