CMS-257 Use new createOrUpdate rpc call in admin
Use new createOrUpdate account rpc call in admin for info. To call the method form the UI, you can use the following pattern:

Admin.lib.RemoteService.account_createOrUpdate( { .. }, function(r) { } );

The profile is not yet supported, so the profile can now be ignored.

The key is not the old "48294879823FB4892793" format, but the new "user:store:name" format. The search panel still uses the old format, but also sends over the new key format (new_key). Use this to pass in to the new rpc method.