CMS-397 New UI: Tile based start menu
We want do redesign the start menu. The design can be created outside of the scopes of existing admin for now. Every application has it's own tile. Each tile is rendered by the application javascript code (parts of it). We only care about the initial design right now and not the "add"/"remove" logic.

Also, use standard HTML and only use ExtJS for what it's worth. Not any facy Ext panels.