CMS-513 Content Types 2.0: Generate form for content edit from content type
Implement dynamic generation of the form to edit content data (Data step in new/edit content in Content Manager app). The fields and layout of the form depends on the content type definition.

This task is only for the UI part, there is a separate task for creating the RPC handler that will return the details of the content type.

    Skip implementation of validation logic: required, immutable, occurrences, validationRegexp.
    The field "customText" does not have anything to do with gui. Use "label" as label text for the field.
    The"helpText" is typically used as information to the user about input. Should be displayed as a tooltip or similar.
    The "inputType" specifies the kind of input in the GUI (ignore the value of "formItemType").