CMS-616 Content types 2.0: Adjustments in content edit
 The New button in Content Manager should display the list of content types, not open the content edit tab as it does now. It's necessary to select a content type for creating the new content.
Avoid opening a new tab when selecting Edit on a content that is already open in a tab. It should work as in Accounts, clicking Edit in a content that is already open in a tab, should change the focus to that tab.
Changing the display name in an edited content and saving should not create a new content.
Show path of the content being edited, where it currently is "/some/content/path".