CMS-818 Live view and info view support in "details pane"
When selecting an item in the content manager app, it automatically opens in the details pane below.

This pane must now support two modes:
"Live View" and "Info View", where Live will be the default mode. This must also be implemented when item is opened as separate "TAB".

Notice however that the toggle modes are slightly different placed and looking, depending on if it is opened in list or in separate tab.