CMS-1027 Content Manager Edit form - Fix array handling for input-type textLine
When an input is listed in the form, it will always begin with a label, then the input will be rendered besides the label. Array handling is thus implemented by the "input" itself (which makes sence since this will also be pluggable, and array handling is different in a google map than textLine etc.).

Problems with current implementation:
* When setting up a form with a textline array with minimum="2" this is handled by rendering two textlines (with label) - which is wrong.
* Also, there is no way of adding more textlines even if maximum is larger (or unlimited).
* Finally if minimum = 0 it is possible to remove the textline field entirely by clicking the x, but no way of getting it back?
Solution:
* Ensure form generation code so that array handling is controlled by the input (not the form renderer)
* Ensure "input type textLine" to support arrays according to attached design
* Ensure "input type textline" so that it cannot be removed entirely according to attached design
NB! The "required number in red" should be rendered by the form renderer, not by the input (fix this also)