Dropdown_fix
Single viewer object was used to display selected option and to update options listed in dropdown grid. When options in dropdown grid are rendered in formatter, viewer gets new object, uses it to create html and converts it to string. Actually DOM structure, which was displayed as selected option, was used to update dropdown options.
I changed DropdownGrid to create new viewer object of the same class that comes from config.