CMS-4549 User Manager - New Group and Role
Changed the behavior of the `New` button, so it will only be enabled,
when the either one and only of the following folders is selected:
Groups, Roles or Users.
Clicking on it will call the NewPrincipalAction and fire
NewPrincipalEvent.
The key of the newly created Principal is automatically generated for
now (we may change this in future).
Created Principal can be saved and edited. (Grid won't be updated.)
Fixed tabs routing.
Fixed Path naming in group/role wizard panels.