CMS-4855 App Header - New Tab design - Browser-like tab design
Discovered and fixed the bug with tabs selection: after the selected tad
was closed, another one wasn't visually selected, as it should.
Updated the TabItem's close button code, as @jvs asked.
Meanwhile fixed the WizardPanel and ItemViewPanel actions key bindings
for all of the ItemViewPanel's actions and WizardPanel's close action,
that were not working for quite a long time (as for the ItemViewPanel).