XP-1382 Component View - Fix keyboard navigation in the tree
The actual fix was the same, as in CMS-3733, which I've solved earlier. There was a conflict between the default slickgrid navigation and the KeyBinding. The new `PageComponentsTreeGrid` class is using the custom Options, so the `enableCellNavigation` was set to `true`, which is wrong. I've set it to `false` and added a comments about the conflict.