XP-4227 Track references embedded in HTML-areas
Made the regexp detect images with correct tags. Simpyfied the regexp a bit since we pretty much now how the result from the HTML-area inserts are formatted, and if it doesnt work here it wont be detected in the portal either, which makes sense. Now, this class and HtmlLinkProcessor should be syncronized and use the same patterns, but I dont have time for that now. Creating a task for that.