Use Content as type parameter of query() + similar #9808
The return type of `contentLib.query()` (and similar functions) should
have the shape `Content<A, 'a'> | Content<B, 'b'>` instead of the
current implementation which is `Content<A | B, 'a' | 'b'>`.

The solution is to take in the whole union of Content as an input
parameter.

(cherry picked from commit 3d6d0ff97f9df5976be4cececc3246b60438bf07)