Add default parameters for AggregationInput and Output #10101
`never` is used as the default type parameter for the `aggregation`
properties in the different QueryParams.

`undefined` is used as the default type parameter for `aggregations`
in different query results. The reason `never` is not used is that
since the property is not optional, we are not able to construct this
object independently of the library functions.